Demo/users/requirements.txt
mongoengine==0.29.1
pymongo==4.14.1
dnspython==2.7.0
djangorestframework==3.15.2